TP3: Fourier Transform
You will use the OpenCV library to perform the following tasks:

1) Fourier Transform:
Utilize the functions provided by OpenCV to calculate the Fourier transform and its inverse. You should:
- Display the original image.
- Calculate the Fourier transform and then compute the corresponding spectrum. Display the spectrum.
- Compute the inverse Fourier transform to obtain the image in the spatial domain. Display the resulting image.

2) Low-pass and High-pass Filters:
Attempt to create low-pass and high-pass filters in the frequency domain to filter images. You can experiment with different cutoff frequency values. To facilitate your work, here are the basic steps:
- Apply the Fourier transform to the image.
- Create filters (high-pass and low-pass) on the real and imaginary parts.
- Apply the inverse Fourier transform to these two parts (imaginary and real).
Try testing with different cutoff frequencies. It is essential to visualize the filters, the obtained results, and analyze them.

3) Filtering sinusoidal noise (optional):
This part is optional. If you complete this section, a bonus will be awarded.