TP4: Isolating Characters – Processing Pipeline
Deadline: 11/01/2023 (Hanoi time)

You will be using the OpenCV library to perform the following task:
Suppose we need to recognize characters in an image. To achieve this, we can separate each of the characters. Your task is to define a processing pipeline with specific techniques at each stage to isolate different characters and then count them.
To do this, several steps are necessary:
- Pre-processing: Images may not be of good quality, so pre-processing is required to enhance image quality and facilitate subsequent steps.
- Character Localization: Various techniques are possible. You can segment images and identify connected components. Alternatively, you can detect closed contours to find letters.
- Post-processing: The same processes need to be applied to all images or to all images in each group. You must create a program that takes an image as input and performs the same processes as on the others.